refactor: move packet-listener into core as all logic is shared on all supported versions
