+++ v1_20_R4/src/main/java/io/th0rgal/oraxen/nms/v1_20_R4/furniture/FurniturePacketManager.java
@@ -3,15 +3,11 @@ package io.th0rgal.oraxen.nms.v1_20_R4.furniture;
 import com.ticxo.modelengine.api.ModelEngineAPI;
 import com.ticxo.modelengine.api.generator.blueprint.ModelBlueprint;
 import io.papermc.paper.math.Position;
-import io.th0rgal.oraxen.OraxenPlugin;
-import io.th0rgal.oraxen.mechanics.MechanicsManager;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.furniture.*;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.furniture.hitbox.BarrierHitbox;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.furniture.hitbox.InteractionHitbox;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.light.LightBlock;
 import io.th0rgal.oraxen.utils.BlockHelpers;
-import io.th0rgal.oraxen.utils.VersionUtil;
-import io.th0rgal.oraxen.utils.logs.Logs;
 import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
 import net.minecraft.network.protocol.game.ClientboundRemoveEntitiesPacket;
 import net.minecraft.network.protocol.game.ClientboundSetEntityDataPacket;
@@ -34,16 +30,6 @@ import org.jetbrains.annotations.NotNull;
 import java.util.*;
 
 public class FurniturePacketManager implements IFurniturePacketManager {
-
-    public FurniturePacketManager() {
-        if (VersionUtil.isPaperServer()) {
-            MechanicsManager.registerListeners(OraxenPlugin.get(), "furniture", new FurniturePacketListener());
-        } else {
-            Logs.logWarning("Seems that your server is a Spigot-server");
-            Logs.logWarning("FurnitureHitboxes will not work due to it relying on Paper-only events");
-            Logs.logWarning("It is heavily recommended to make the upgrade to Paper");
-        }
-    }
 
     private final int INTERACTION_WIDTH_ID = 8;
     private final int INTERACTION_HEIGHT_ID = 9;
